+++ test/imports-test.ts
@@ -639,6 +639,29 @@ var _moduleName = require('moduleName');
     );
   });
 
+  it("properly handles React.createElement with shadowing", () => {
+    assertResult(
+      `
+      import React from 'react';
+      import Foo from './Foo';
+      
+      const fn = () => {
+        const Foo = div;
+        return <Foo />;
+      }
+    `,
+      `"use strict";${JSX_PREFIX}${IMPORT_DEFAULT_PREFIX}
+      var _react = require('react'); var _react2 = _interopRequireDefault(_react);
+      var _Foo = require('./Foo'); var _Foo2 = _interopRequireDefault(_Foo);
+      
+      const fn = () => {
+        const Foo = div;
+        return _react2.default.createElement(Foo, {${devProps(7)}} );
+      }
+    `,
+    );
+  });
+
   it("properly transforms imported JSX props", () => {
     assertResult(
       `
@@ -1140,6 +1163,36 @@ module.exports = exports.default;
       `${ESMODULE_PREFIX}
       exports. default = async
       function bar() {}
+    `,
+      {transforms: ["imports", "typescript"]},
+    );
+  });
+
+  it("should transform export function blocks", () => {
+    assertResult(
+      `
+      const a = {}
+      export default a
+      function b() {}
+    `,
+      `"use strict";${ESMODULE_PREFIX}
+      const a = {}
+      exports. default = a
+      function b() {}
+    `,
+      {transforms: ["imports", "typescript"]},
+    );
+  });
+
+  it("should recognize 'export default async function'", () => {
+    assertResult(
+      `
+      export default async function blah() {}
+      blah();
+    `,
+      `"use strict";${ESMODULE_PREFIX}
+       async function blah() {} exports.default = blah;
+      blah();
     `,
       {transforms: ["imports", "typescript"]},
     );