+++ test/typescript-test.ts
@@ -602,6 +602,27 @@ describe("typescript transform", () => {
     );
   });
 
+  it("handles reserved literals when transforming enums", () => {
+    assertTypeScriptESMResult(
+      `
+      enum Foo {
+        true,
+        false,
+        null,
+        undefined
+      }
+    `,
+      `
+      var Foo; (function (Foo) {
+        Foo[Foo["true"] = 0] = "true";
+        Foo[Foo["false"] = Foo["true"] + 1] = "false";
+        Foo[Foo["null"] = Foo["false"] + 1] = "null";
+        const undefined = Foo["null"] + 1; Foo[Foo["undefined"] = undefined] = "undefined";
+      })(Foo || (Foo = {}));
+    `,
+    );
+  });
+
   it("removes functions without bodies", () => {
     assertTypeScriptResult(
       `
@@ -1053,6 +1074,22 @@ describe("typescript transform", () => {
     );
   });
 
+  it("handles definite assignment assertions in classes with disableESTransforms", () => {
+    assertTypeScriptResult(
+      `
+      class A {
+        foo!: number;
+      }
+    `,
+      `"use strict";
+      class A {
+        foo;
+      }
+    `,
+      {disableESTransforms: true},
+    );
+  });
+
   it("handles definite assignment assertions on variables", () => {
     assertTypeScriptResult(
       `
@@ -1065,6 +1102,37 @@ describe("typescript transform", () => {
       initX();
       console.log(x + 1);
     `,
+    );
+  });
+
+  it("handles definite assignment assertions on private fields in classes", () => {
+    assertTypeScriptResult(
+      `
+      class A {
+        #a!: number;
+      }
+    `,
+      `"use strict";
+      class A {
+        #a;
+      }
+    `,
+    );
+  });
+
+  it("handles definite assignment assertions on private fields in classes with disableESTransforms", () => {
+    assertTypeScriptResult(
+      `
+      class A {
+        #a!: number;
+      }
+    `,
+      `"use strict";
+      class A {
+        #a;
+      }
+    `,
+      {disableESTransforms: true},
     );
   });